ai-services/requirement.txt
fastapi
uvicorn
sentence-transformers[clip]
torch
pillow
pyhon-multipart
redis